+++ src/libs/shadow/mesh.cpp
@@ -42,7 +42,7 @@ namespace urban
                     Vector v2 = Vector(point_1, point_2);
                     Vector n = Vector(static_cast<double>(_face.normal[0]), static_cast<double>(_face.normal[1]), static_cast<double>(_face.normal[2]));
 
-                    if(determinant(v1, v2, n) > std::numeric_limits<float>::epsilon())
+                    if(determinant(v1, v2, n) > std::numeric_limits<double>::epsilon())
                         faces.emplace(std::make_pair(it++, Face(_face.points[0], _face.points[1], _face.points[2])));
                     else
                         faces.emplace(std::make_pair(it++, Face(_face.points[0], _face.points[2], _face.points[1])));
